chore(team): split mailbox service channels
